requirements.txt
jieba
tqdm
tensorboardX
torch
pytorch-nlp
gensim
pandas
sklearn